Refactored models and layers
