+++ neuralop/layers/einsum_utils.py
