+++ neuralop/layers/spectral_convolution.py
