+++ neuralop/layers/tests/test_spectral_convolution.py
