+++ neuralop/models/uno.py
@@ -2,15 +2,11 @@ import torch.nn as nn
 import torch.nn.functional as F
 from functools import partialmethod
 import torch
-from .mlp import MLP
-from .spectral_convolution import SpectralConv3d, SpectralConv2d, SpectralConv1d
-from .spectral_convolution import SpectralConv
-from .skip_connections import skip_connection
-from .padding import DomainPadding
-from .fno_block import FNOBlocks,resample
-from .fno import partialclass
-import numpy as np
-#this will be merged with the neural operator UNO
+from ..layers.mlp import MLP
+from ..layers.spectral_convolution import SpectralConv
+from ..layers.skip_connections import skip_connection
+from ..layers.padding import DomainPadding
+from ..layers.fno_block import FNOBlocks,resample
 
 class UNO(nn.Module):
     """
